requirements.txt
flask>=3.0
python-dotenv>=1.0
httpx>=0.27
spacy>=3.7
# fr_core_news_sm : python -m spacy download fr_core_news_sm